requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.0
pydantic-settings==2.1.0
selectolax==0.3.21
yake==0.4.8
keybert==0.8.4
sentence-transformers==2.3.1
scikit-learn==1.4.0
nltk==3.8.1
python-dotenv==1.0.0
python-multipart==0.0.6

